backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
azure-identity>=1.15.0
azure-core>=1.29.0
azure-cognitiveservices-speech>=1.35.0
azure-ai-translation-text>=1.0.0
aiofiles>=23.2.0
